renamed app & clean-up
