+++ Frontend/src/components/Home.tsx
@@ -2,9 +2,9 @@ import React from 'react'
 
 export const Home: React.FC = () => {
 	return (
-		<>
-			<h1>Home page</h1>
+		<div className="flex justify-center align-middle h-100">
+			<span className="text-xl py-10">Home page</span>
 			<br />
-		</>
+		</div>
 	)
 }
