+++ Frontend/src/components/Router.tsx
@@ -1,9 +1,5 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
-// import Home from './Home';
-// import Navbar from './Navbar';
-// import SignIn from './SignIn';
-// import SignOutBtn from './SignOutBtn';
-// import Statistics from './stats/Statistics';
+import { Home, Navbar, SignIn, SignOutBtn, Statistics } from '.'
 import React, { useContext } from 'react'
 import { AuthContext } from '../contexts'
 
@@ -12,18 +8,18 @@ export const Router: React.FC = () => {
 	return (
 		<>
 			<BrowserRouter>
-				{/* <Navbar /> */}
+				<Navbar />
 				<Routes>
-					<Route path="/" element={<p>HELLO THERE</p>} />
-					{/* <Route exact path="/" element={<Home />} />
-					<Route exact path="/about" element={<div>About page</div>} />
-					<Route exact path="*" element={<Home />} /> */}
+					<Route path="*" element={<Home />} />
+					<Route path="/" element={<Home />} />
 					{signedIn === false ? (
-						<>{/* <Route exact path="/sign-in" element={<SignIn />} /> */}</>
+						<>
+							<Route path="/sign-in" element={<SignIn />} />{' '}
+						</>
 					) : (
 						<>
-							{/* <Route exact path="/stats" element={<Statistics />} />
-							<Route exact path="/sign-out" element={<SignOutBtn />} /> */}
+							<Route path="/stats" element={<Statistics />} />
+							<Route path="/sign-out" element={<SignOutBtn />} />
 						</>
 					)}
 				</Routes>
