--- TEST_APP/rehtracker-frontend/package.json
@@ -1,57 +1,0 @@
-{
-	"name": "rehtracker-frontend",
-	"version": "0.1.0",
-	"private": true,
-	"dependencies": {
-		"@testing-library/jest-dom": "^5.17.0",
-		"@testing-library/react": "^13.4.0",
-		"@testing-library/user-event": "^13.5.0",
-		"@types/date-fns": "^2.6.0",
-		"@types/jest": "^29.5.4",
-		"@types/node": "^20.6.0",
-		"@types/react": "^18.2.21",
-		"@types/react-dom": "^18.2.7",
-		"@types/react-router-dom": "^5.3.3",
-		"@types/styled-components": "^5.1.27",
-		"axios": "^1.5.0",
-		"date-fns": "^2.30.0",
-		"react": "^18.2.0",
-		"react-day-picker": "^8.8.2",
-		"react-dom": "^18.2.0",
-		"react-icons": "^4.11.0",
-		"react-router-dom": "^6.16.0",
-		"react-scripts": "5.0.1",
-		"styled-components": "^6.0.8",
-		"typescript": "^5.2.2",
-		"web-vitals": "^2.1.4"
-	},
-	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject"
-	},
-	"eslintConfig": {
-		"extends": [
-			"react-app",
-			"react-app/jest"
-		]
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	},
-	"devDependencies": {
-		"autoprefixer": "^10.4.15",
-		"postcss": "^8.4.29",
-		"tailwindcss": "^3.3.3"
-	}
-}
