--- TEST_APP/rehtracker-frontend/src/App.css
@@ -1,17 +1,0 @@
-.App {
-	text-align: center;
-	height: 100%;
-}
-
-.App-link {
-	color: #61dafb;
-}
-
-html,
-body {
-	height: 100%;
-}
-
-#root {
-	height: 100%;
-}
