--- TEST_APP/rehtracker-frontend/src/components/Home.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-export const Home: React.FC = () => {
-	return (
-		<div className="flex justify-center align-middle h-100">
-			<span className="text-xl py-10">Home page</span>
-			<br />
-		</div>
-	)
-}
