--- TEST_APP/rehtracker-frontend/src/components/Router.tsx
@@ -1,29 +1,0 @@
-import { BrowserRouter, Routes, Route } from 'react-router-dom'
-import { Home, Navbar, SignIn, SignOutBtn, Statistics } from '.'
-import React, { useContext } from 'react'
-import { AuthContext } from '../contexts'
-
-export const Router: React.FC = () => {
-	const { signedIn } = useContext(AuthContext)
-	return (
-		<>
-			<BrowserRouter>
-				<Navbar />
-				<Routes>
-					<Route path="*" element={<Home />} />
-					<Route path="/" element={<Home />} />
-					{signedIn === false ? (
-						<>
-							<Route path="/sign-in" element={<SignIn />} />{' '}
-						</>
-					) : (
-						<>
-							<Route path="/stats" element={<Statistics />} />
-							<Route path="/sign-out" element={<SignOutBtn />} />
-						</>
-					)}
-				</Routes>
-			</BrowserRouter>
-		</>
-	)
-}
