--- TEST_APP/rehtracker-frontend/src/components/index.tsx
@@ -1,7 +1,0 @@
-export * from './Router'
-export * from './Home'
-export * from './Navbar'
-export * from './SignIn'
-export * from './SignOutBtn'
-export * from './SignUp'
-export * from './stats'
