--- TEST_APP/rehtracker-frontend/src/components/stats/index.tsx
@@ -1,3 +1,0 @@
-export * from './ExerChart'
-export * from './offlineData'
-export * from './Statistics'
