--- TEST_APP/rehtracker-frontend/src/contexts/index.tsx
@@ -1,1 +1,0 @@
-export * from './AuthContext'
