--- TEST_APP/rehtracker-frontend/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"extends": "../tsconfig.json",
-	"compilerOptions": {
-		"rootDir": "./src/",
-		"outDir": "./dist/",
-		"lib": ["dom", "dom.iterable", "esnext"],
-		"forceConsistentCasingInFileNames": true,
-		"module": "esnext",
-		"resolveJsonModule": true,
-		"isolatedModules": true,
-		"noEmit": true
-	},
-	"include": ["**/*.ts", "**/*.tsx"],
-	"exclude": ["src/**/*.test.ts", "node_modules"]
-}
