+++ frontend/src/App.css
@@ -1,71 +1,17 @@
-@import url('https://fonts.googleapis.com/css2?family=Sora');
-
-:root {
-  --rdp-cell-size: 35px;
-  --rdp-accent-color: #06acd666;
-  --rdp-background-color: #706f6f51;
-  /* Switch to dark colors for dark themes */
-  /* --rdp-accent-color-dark: #3003e1; */
-  /* --rdp-background-color-dark: #180270; */
-  /* Outline border for focused elements */
-  /* --rdp-outline: 2px solid var(--rdp-accent-color); */
-  /* Outline border for focused and selected elements */
-  --rdp-outline-selected: 0px solid #06d6a0;
-  /* --rdp-align-items: 'center';
-  --rdp-flex: 1;
-  --rdp-justify-content: 'center'; */
-  
-}
-
-.rdp{
-  align-items: center;
-  justify-content: center;
-  margin-top: 45pt;
-}
-
-body, html {
-  padding: 0%;
-  background-color: rgba(255, 254, 243, 0.938);
-  font-family: 'Sora', sans-serif;
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-}
-
 .App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+	text-align: center;
+	height: 100%;
 }
 
 .App-link {
-  color: #61dafb;
+	color: #61dafb;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+html,
+body {
+	height: 100%;
+}
+
+#root {
+	height: 100%;
 }
